updated the upload controller ready to start handling urls. added upload-method and url pages
