+++ src/controllers/uploadUrlController.js
@@ -1,0 +1,7 @@
+import UploadController from './uploadController.js'
+
+class UploadUrlController extends UploadController {
+
+}
+
+export default UploadUrlController
