+++ src/views/start.html
@@ -30,7 +30,7 @@
   
   {{ govukButton({
     text: "Start now",
-    href: "/dataset",
+    href: "/upload-method",
     isStartButton: true
   }) }}
